Refactor Vietnamese text normalization and optimize performance
- Split `text_norm.py` into smaller sub-modules for better maintainability.
- Optimize measurement/currency expansion using a consolidated single-pass regex.
- Consolidate acronym exception replacements into a single pass.
- Remove unused regexes in `symbols.py`.
- Maintain backward compatibility in `text_norm.py`.
